Remove DisallowedUriList and AllowedUriList in favour of using UriList for both allowing and disallowing URIs
